requirements.txt
Flask
firebase-admin
python-dotenv
requests
imagekitio==4.2.0
fpdf
pillow
tqdm